nix: add lint, test-uring packages; fix clippy errors
- Add lint and lint-full packages to flake.nix using nix run
- Add test-uring package for lio-uring tests
- Add linuxHeaders to test packages for Linux builds
- Add libiconv detection in test-ffi.sh for macOS linking
- Fix clippy undocumented_unsafe_blocks errors
- Fix unnecessary_cast warnings with allow attributes
- Fix unnecessary_safety_comment errors
- Remove needless return statements

+++ lio/src/backend/impls/pollingv2/os/epoll.rs
@@ -116,7 +116,7 @@ impl OsPoller {
     }
     // Note: EPOLLHUP and EPOLLERR are always reported by the kernel regardless of registration
 
-    let mut event = libc::epoll_event { events, u64: key as u64 };
+    let mut event = libc::epoll_event { events, u64: key };
 
     syscall!(epoll_ctl(
       self.epoll_fd.as_raw_fd(),
@@ -173,7 +173,7 @@ impl ReadinessPoll for OsPoller {
     events |= libc::EPOLLONESHOT as u32;
     // Note: EPOLLHUP and EPOLLERR are always reported by the kernel regardless of registration
 
-    let mut event = libc::epoll_event { events, u64: key as u64 };
+    let mut event = libc::epoll_event { events, u64: key };
 
     syscall!(epoll_ctl(
       self.epoll_fd.as_raw_fd(),
